Module burn_wgpu::kernel

Kernel module

Modules

conv

Convolution kernels

matmul

Matmul kernels

pool

Pooling kernels

prng

Pseudo-random number generator kernels

reduce

Reduction algorithms

Macros

binary_elemwise

Creates a binary elementwise kernel.

binary_elemwise_inplace

Creates a binary elementwise inplace kernel.

kernel_wgsl

Generates kernel source code by replacing some information using templating.

unary

Creates a unary kernel.

unary_inplace

Creates a unary inplace kernel.

unary_scalar

Creates a unary scalar kernel.

unary_scalar_inplace

Creates a unary scalar inplace kernel.

Structs

BinaryElemwiseInplaceRaw

Generated kernel from wgsl file.

BinaryElemwiseRaw

Generated kernel from wgsl file.

CastRaw

Generated kernel from wgsl file.

ContiguousRaw

Generated kernel from wgsl file.

DynamicKernelSettings

Generate kernel source code by replacing some information using templating.

KernelSettings

Generates kernel source code by replacing some information using templating.

SourceTemplate

Kernel source code abstraction allowing for templating.

UnaryInplaceRaw

Generated kernel from wgsl file.

UnaryRaw

Generated kernel from wgsl file.

UnaryScalarInplaceRaw

Generated kernel from wgsl file.

UnaryScalarRaw

Generated kernel from wgsl file.

Traits

DynamicKernelSource

Dynamic wgpu kernel to create a source template.

StaticKernelSource

Static wgpu kernel to create a source template.

Functions

binary_elemwise

Execute a binary kernel using the provided WORKGROUP.

binary_elemwise_default

Execute a binary kernel using the default settings.

binary_elemwise_inplace

Execute a binary inplace kernel using the provided WORKGROUP.

binary_elemwise_inplace_default

Execute a binary inplace kernel using the default settings.

build_info

Create a vector containing the dimension, strides and shape of tensors.

build_info_dyn

Similar to build info but with dynamic rank.

cast

Cast a tensor to the given element type.

into_contiguous

Make a wgpu tensor contiguous.

into_contiguous_dyn

Similar to into contiguous but with dynamic rank.

unary

Execute a unary kernel using the provided WORKGROUP.

unary_default

Execute a unary kernel using the default settings.

unary_inplace

Execute a unary inplace kernel using the provided WORKGROUP.

unary_inplace_default

Execute a unary inplace kernel using the default settings.

unary_scalar

Execute a unary scalar kernel using the provided WORKGROUP.

unary_scalar_default

Execute a unary scalar kernel using the default settings.

unary_scalar_inplace

Execute a unary scalar inplace kernel using the provided WORKGROUP.

unary_scalar_inplace_default

Execute a unary scalar inplace kernel using the default settings.